US - 13 - frontend (fix some routing bugs)
